// src/app/core/interfaces/worker-get.interface.ts
export interface WorkerGet {
    id: string;
    photo: string;
    stateId: string;
    stateName: string;
    typeId: string;
    typeName: string;
    authid: string;
    email: string;
    datecreate: string;
    typedocumentId: string;
    typedocumentName: string;
    numerodoc: string;
    name: string;
    firstname: string;
    lastname: string;
    address: string;
    phone: string;
    fullname: string;
}